lablu_system/requirements.txt
PyQt6>=6.5.0
reportlab>=4.0.0
Pillow>=10.0.0